Remove unused code from pkg/icons
